tests/wpt/css-tests/css-flexbox-1_dev/html/flexbox_computedstyle_align-self-stretch.htm

<!DOCTYPE html>
<html><head><title>flexbox | computed style | align-self: stretch</title>
<link href="http://opera.com" rel="author" title="Opera Software">
<link href="http://www.w3.org/TR/css-flexbox-1/#propdef-align-self" rel="help">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags">
<style>
body {
	color: red;
	align-self: stretch;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("align-self"),
	          "stretch");
});
</script>
</body></html>
